converted S'tack from int/float to long/double (not BigInt cuz I dunno how to use that)

+++ src/net/ncplanner/plannerator/planner/s_tack/object/StackFloat.java
@@ -1,6 +1,6 @@
 package net.ncplanner.plannerator.planner.s_tack.object;
 public class StackFloat extends StackNumber{
-    public StackFloat(float value){
+    public StackFloat(double value){
         super(value);
     }
     @Override
@@ -8,11 +8,11 @@ public class StackFloat extends StackNumber{
         return Type.FLOAT;
     }
     @Override
-    public Float getValue(){
-        return value.floatValue();
+    public Double getValue(){
+        return value.doubleValue();
     }
     @Override
     public StackObject duplicate(){
-        return new StackFloat(value.floatValue());
+        return new StackFloat(value.doubleValue());
     }
 }
