+++ src/net/ncplanner/plannerator/planner/s_tack/token/operator/PowerOperator.java
@@ -13,9 +13,9 @@ public class PowerOperator extends Operator{
     @Override
     public StackObject evaluate(StackObject v1, StackObject v2){
         if(v1.getBaseType()==StackObject.Type.INT&&v2.getBaseType()==StackObject.Type.INT){
-            return new StackInt((int)Math.pow(v1.asInt().getValue(), v2.asInt().getValue()));
+            return new StackInt((long)Math.pow(v1.asInt().getValue(), v2.asInt().getValue()));
         }else{
-            return new StackFloat((float)Math.pow(v1.asNumber().getValue().floatValue(), v2.asNumber().getValue().floatValue()));
+            return new StackFloat(Math.pow(v1.asNumber().getValue().doubleValue(), v2.asNumber().getValue().doubleValue()));
         }
     }
 }
